Solution to Toll Plaza Management System

+++ Toll Plaza Management System@@67877d61dc107486ad5c2750/code.c
@@ -5,10 +5,10 @@ int main(){
     scanf("%d",&n);
     char number[n][10];
     char type[n][10];
-    int tax[n];
+    float tax[n];
     float car_tax=0,truck_tax=0,bike_tax=0;
     for(int i=1;i<=n;i++){
-        scanf("%s %s %d",number[i],type[i],&tax[i]);
+        scanf("%s %s %f",number[i],type[i],&tax[i]);
         if(type[i]=="Car"){
             car_tax+=tax[i];
         }else if(type[i]=="Truck"){
